+++ CHANGELOG.md
@@ -7,6 +7,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Fixed
+- Исправлен путь к main файлу в package.json для корректного импорта пакета
+
 ## [0.8.2] - 2025-04-16
 
 ### Fixed